refactor(binding-opcua): add ContentSerdes for OPC UA binding
