+++ packages/browser-bundle/package.json
@@ -19,8 +19,7 @@
     "browserify": "16.2.3",
     "tinyify": "2.5.0"
   },
-  "dependencies": {
-  },
+  "dependencies": {},
   "scripts": {
     "build": "browserify index.js --plugin tinyify -o dist/wot-bundle.min.js"
   }
