+++ packages/core/package.json
@@ -15,21 +15,23 @@
   "types": "dist/core.d.ts",
   "devDependencies": {
     "@types/chai": "4.0.4",
-    "@types/node": "10.9.4",
+    "@types/node": "^10.17.14",
     "chai": "4.1.2",
     "mocha": "3.5.3",
     "mocha-typescript": "1.1.8",
     "ts-node": "3.3.0",
-    "typescript": "2.9.2",
+    "typescript": "3.6.3",
     "typescript-standard": "0.3.30"
   },
   "dependencies": {
-    "wot-typescript-definitions": "0.7.1-SNAPSHOT.1",
     "@node-wot/td-tools": "0.7.0-SNAPSHOT.5",
-    "uuid": "3.3.3",
+    "@types/uritemplate": "0.3.4",
+    "node-opcua-client": "^2.1.8",
+    "xml-writer": "^1.7.0",
     "rxjs": "5.4.3",
     "uritemplate": "0.3.4",
-    "@types/uritemplate": "0.3.4"
+    "uuid": "3.3.3",
+    "wot-typescript-definitions": "0.7.1-SNAPSHOT.1"
   },
   "scripts": {
     "build": "tsc",
